# Implementation Summary
-------------------------------
With removal of the frontend file picker, enterprise users are unable to create a dataset. Remove the need for picking the csv by creating a dataset directly after uploading a zip containing one.

Highlights of rygg changes:
* Added POST /datasets/create_from_upload, which works like POST /datasets/create_from_remote, except that it receives an uploaded zip
* Removed the /upload endpoints so we can better categorize what's being uploaded by uploading files to datasets or models
* Removed obsolete modeldirectories endpoints
* Removed json_models endpoints so that model jsons can be saved to the models/<id>/save_json endpoint
* Keep track of the dataset's root directory instead of just the csv. This allows rygg to delete the dataset cleanly when that's requested

Highlights of frontend changes:
* Adapt to changes in the rygg interface
* add an Esc-key handler in select-model-modal
* refactor autoPopulateName in select-model-modal in preparation for moving that logic to rygg.
* Add initial change for uploading dataset zips to loadDataset in projects.vue. This is a place that needs attention from the frontend team, since I just added a sleep to make it wait for rygg to unzip. It needs to poll the task's status and probably show a progress bar.

This PR is phase 1: make the required changes in rygg and get the frontend working for local installations.

(Phase 2 will be making the frontend work for docker installations.)

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Refined tests

# Please also:
- [X] Include any new files in included_files.txt
- [X] Update the story on the board

Related work items: #2477

+++ rygg/rygg/files/views/file_view.py
@@ -1,19 +1,15 @@
-from contextlib import contextmanager
 from django_http_exceptions import HTTPExceptions
-from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 from rest_framework.generics import RetrieveDestroyAPIView
 from rest_framework.response import Response
-import chardet
+import itertools
 import os
 
-from rygg.files.utils.file_choosing import open_file_dialog, open_saveas_dialog
-
 from rygg.files.paths import PathNotAvailable
-from rygg.files.utils.file_choosing import open_file_dialog
+from rygg.files.utils.file_choosing import open_file_dialog, open_saveas_dialog
+from rygg.files.utils.file import get_text_lines
 from rygg.files.views.util import (
     get_path_param,
-    get_required_param,
     get_optional_param,
     json_response,
     make_path_response,
@@ -34,44 +30,17 @@ def full_path(request):
 
 class FileView(RetrieveDestroyAPIView):
     def get(self, request, format=None):
+        if IS_CONTAINERIZED:
+            raise HTTPExceptions.BAD_REQUEST.with_content("not supported in docker installations")
         path = full_path(request)
         return make_path_response(path)
 
     def delete(self, request, format=None):
+        if IS_CONTAINERIZED:
+            raise HTTPExceptions.BAD_REQUEST.with_content("not supported in docker installations")
         path = full_path(request)
         os.remove(path)
         return make_path_response(path)
-
-@api_view(["GET"])
-def get_file_content(request):
-    path = full_path(request)
-    num_rows = get_optional_param(request, "num_rows", 4) #5 rows, 0 indexed
-    return make_file_content_response(num_rows, path)
-
-
-def make_file_content_response(num_rows, file_path):
-
-    row_count = 0
-    row_contents = []
-    encoding=get_encoding(file_path)
-
-    try:
-        with open(file_path, 'r', encoding=encoding, errors='backslashreplace') as reader:
-            line = reader.readline()
-            while line != '' and row_count < num_rows:
-                row_contents.append(line)
-                line = reader.readline()
-                row_count += 1
-    except Exception as err:
-        raise Exception('Error when reading file contents')
-
-    return json_response({"file_contents": row_contents})
-
-def get_encoding(file_path):
-    with open(file_path, 'rb') as file:
-        raw = file.read(32) # at most 32 bytes are returned
-        return chardet.detect(raw)['encoding']
-
 
 import json
 def get_file_types_from_request(request):
